Tests undertaken pass
Write Data to a file so we know where it is.
Read the data back and initialise a ScoreSummary Object.
Added a task to the game so that we could check if the time update worked. Several problems occurred but are not all passing.
Refactored by creating a method or function that simplifies the understanding of the C# code
Ran regression tests and all still pass!